ensuring async calls to fetch jenkins client, moving calls to create client outside constructors using factory, placing jenkins namespace at the root of Core project for clarity since it's used outside the engine.
