+++ src/Core/Azure/DefaultManagedIdentityService.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
+using Modm.Azure.Model;
 
 namespace Modm.Azure
 {
@@ -103,4 +104,3 @@ namespace Modm.Azure
         }
     }
 }
-
