--- src/Core/Azure/KuduApi.cs
@@ -1,43 +1,0 @@
-﻿using System;
-using Azure.Identity;
-using Azure.Core;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.IO;
-using System.Threading.Tasks;
-
-namespace Modm.Azure
-{
-    public class KuduApi
-    {
-        private readonly HttpClient client;
-        private readonly string functionAppName;
-        private readonly DefaultAzureCredential credential;
-
-        public KuduApi(string functionAppName, HttpClient client)
-        {
-            this.functionAppName = functionAppName;
-            this.client = client;
-            this.credential = new DefaultAzureCredential();
-        }
-
-        public async Task DeployZipAsync(string zipFilePath)
-        {
-            var baseUrl = $"https://{functionAppName}.scm.azurewebsites.net/api/zipdeploy/";
-
-            using var request = new HttpRequestMessage(HttpMethod.Post, baseUrl);
-
-            var token = await credential.GetTokenAsync(new TokenRequestContext(new[] { "https://management.azure.com/.default" }));
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", token.Token);
-
-            using var fileStream = new FileStream(zipFilePath, FileMode.Open);
-            using var streamContent = new StreamContent(fileStream);
-
-            request.Content = streamContent;
-
-            var response = await this.client.SendAsync(request);
-            response.EnsureSuccessStatusCode();
-        }
-    }
-}
-
