+++ src/Core/Deployments/DefaultDeploymentRepository.cs
@@ -1,14 +1,10 @@
-﻿using System;
-using System.Threading;
-using JenkinsNET.Models;
-using Microsoft.Extensions.Logging;
-using Modm.Engine.Jenkins.Client;
+﻿using Microsoft.Extensions.Logging;
+using Modm.Jenkins.Client;
 
 namespace Modm.Deployments
 {
     public class DefaultDeploymentRepository : IDeploymentRepository
 	{
-        private IJenkinsClient client;
         private readonly JenkinsClientFactory clientFactory;
         private readonly DeploymentFile file;
         private readonly ILogger<DefaultDeploymentRepository> logger;
@@ -16,32 +12,22 @@ namespace Modm.Deployments
         public DefaultDeploymentRepository(JenkinsClientFactory clientFactory, DeploymentFile file, ILogger<DefaultDeploymentRepository> logger)
 		{
             this.clientFactory = clientFactory;
-            this.client = clientFactory.Create().GetAwaiter().GetResult();
             this.file = file;
             this.logger = logger;
         }
 
-        protected IJenkinsClient JenkinsClient
-        {
-            get
-            {
-                if (this.client == null)
-                {
-                    this.client = this.clientFactory.Create().GetAwaiter().GetResult();
-                }
-                return this.client;
-            }
-        }
-
         public async Task<Deployment> Get(CancellationToken cancellationToken = default)
         {
+            logger.LogTrace("Fetching deployment information");
+
+            using var client = await clientFactory.Create();
+
             var deployment = await file.Read(cancellationToken);
             deployment.Status = await client.GetBuildStatus(deployment.Definition.DeploymentType, deployment.Id);
             deployment.IsStartable = await client.IsJobRunningOrWasAlreadyQueued(deployment.Definition.DeploymentType);
 
             return deployment;
         }
-
     }
 }
